+++ api/src/lib.rs
@@ -24,7 +24,7 @@ pub(crate) type ApiResult<T> = Result<T, ApiError>;
 pub use self::error::ApiError;
 // todo(refactor): unite payloads
 // export payloads
-pub use self::{auth::CredentialsPayload, routes::users::*};
+pub use self::routes::users::*;
 
 pub async fn app(pool: DbPool) -> ApiResult<Router> {
   let session_layer = create_migrate_session_layer(pool.clone()).await?;
