+++ api/src/routes/openapi.rs
@@ -27,7 +27,7 @@ use utoipa_rapidoc::RapiDoc;
 use utoipauto::utoipauto;
 
 use super::users::{delete::*, get::*, post::*, put::*, *};
-use crate::{auth::CredentialsPayload, error::ApiError};
+use crate::error::ApiError;
 
 /// router fragment supplying OpenAPI documentation and ui routes
 /// View rapidoc documentation page at: http://localhost:3000/docs/rapidoc
